feat: add results page (PictureQuiz)
