+++ art-quiz/js/AuthorQuiz.js
@@ -102,6 +102,71 @@ class AuthorQuiz extends Quiz {
     
     return questionView;
   }
+
+  saveResult() {
+    const categoryNum = this.categoryNum;
+    const result = this.result;
+
+    function setLocalStorage() {
+      localStorage.setItem(`result${categoryNum}`, result);
+    }
+    window.addEventListener('beforeunload', setLocalStorage)
+  }
+
+  getResultView() {
+    const results = this.getResult();
+    const categoryNum = this.categoryNum;
+
+    const resultView = document.createElement('div');
+    resultView.classList.add('result');
+
+    const resultTitle = document.createElement('p');
+    resultTitle.classList.add('result__title');
+    resultTitle.textContent = `Категория ${categoryNum}. Раунд завершён`;
+    resultView.append(resultTitle);
+
+    const resultBtn = document.createElement('button');
+    resultBtn.classList.add('result__btn');
+    resultBtn.textContent = 'Результаты';
+    resultView.append(resultBtn);
+
+    resultBtn.addEventListener('click', () => {
+      resultView.textContent = '';
+
+      const resultTitle = document.createElement('p');
+      resultTitle.classList.add('result__title');
+      resultTitle.textContent = `Результат: ${results}/10`;
+      resultView.append(resultTitle);
+
+      const resultCont = document.createElement('div');
+      resultCont.classList.add('result__cont');
+
+      this.result.forEach((isCorrect, ind) => {
+        let imageNum = (categoryNum - 1) * 10 + (ind + 1);
+
+        const resultPicture = document.createElement('img');
+        resultPicture.classList.add('result__item');
+        resultPicture.src = `https://raw.githubusercontent.com/kxzws/image-data/master/img/${imageNum}.jpg`;
+        resultPicture.alt = 'picture: quiz answers';
+
+        if (!isCorrect) resultPicture.classList.add('incorrect');
+
+        resultPicture.addEventListener('click', () => {
+          modalPicture.src = `https://raw.githubusercontent.com/kxzws/image-data/master/img/${imageNum}.jpg`;
+          modalAuthor.textContent = images[imageNum].author;
+          modalName.textContent = images[imageNum].name;
+          modalYear.textContent = images[imageNum].year;
+          toggleModal(isCorrect);
+          document.querySelector('.here-is-button').textContent = '';
+        })
+        
+        resultCont.append(resultPicture);
+      });
+      resultView.append(resultCont);
+    });
+
+    return resultView;
+  }
 }
 
 export default AuthorQuiz;
